requirements.txt
tensorflow==2.20.0
keras==3.10.0  # مش 3.11
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch==2.1.0 --index-url https://download.pytorch.org/whl/cpu
transformers==4.36.0
Pillow==10.1.0
numpy==1.26.0
gdown==5.1.0
typing-extensions==4.8.0

